Split router, controller and rest api
This is the first step towards re-design that would allow
to use both HTTP and WebSocket Handlers.

Currently REST API handlers, router and controller exist as a
single entity. The new architecture looks like:

Router --> Controller --> HTTP response
               |
               v
            Storage
               |
                --> PerfDB
                --> MongoDB

So with this change it is possible to split HTTP response into
HTTP response and WebSocket response depending on client protocol.